+++ projects/github-actions/repo-gardening/src/index.js
@@ -11,18 +11,11 @@ const notifyDesign = require( './tasks/notify-design' );
 const notifyEditorial = require( './tasks/notify-editorial' );
 const replyToCustomersReminder = require( './tasks/reply-to-customers-reminder' );
 const triageIssues = require( './tasks/triage-issues' );
-const updateBoard = require( './tasks/update-board' );
-const wpcomCommitReminder = require( './tasks/wpcom-commit-reminder' );
 const debug = require( './utils/debug' );
 const ifNotClosed = require( './utils/if-not-closed' );
 const ifNotFork = require( './utils/if-not-fork' );
 
 const automations = [
-	{
-		event: 'issues',
-		action: [ 'labeled', 'opened' ],
-		task: updateBoard,
-	},
 	{
 		event: 'pull_request_target',
 		action: [ 'opened', 'synchronize', 'edited' ],
@@ -62,10 +55,6 @@ const automations = [
 		event: 'pull_request_target',
 		action: [ 'labeled' ],
 		task: ifNotClosed( notifyEditorial ),
-	},
-	{
-		event: 'push',
-		task: wpcomCommitReminder,
 	},
 	{
 		event: 'pull_request_target',