+++ projects/js-packages/social-logos/src/react/example.tsx
@@ -68,17 +68,27 @@ function SocialLogosExample() {
 			<div className="display-control-group">
 				<div className="display-control">
 					<h4>Small icons</h4>
-					{ /* eslint-disable-next-line jsx-a11y/label-has-associated-control -- https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/issues/869 */ }
-					<label className="switch">
-						<input type="checkbox" onChange={ handleSmallIconsToggle } checked={ useSmallIcons } />
+					{ /* eslint-disable-next-line jsx-a11y/label-has-associated-control -- https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/issues/578 */ }
+					<label className="switch" htmlFor="useSmallIcons">
+						<input
+							id="useSmallIcons"
+							type="checkbox"
+							onChange={ handleSmallIconsToggle }
+							checked={ useSmallIcons }
+						/>
 						<span className="handle"></span>
 					</label>
 				</div>
 				<div className="display-control">
 					<h4>Icon names</h4>
-					{ /* eslint-disable-next-line jsx-a11y/label-has-associated-control -- https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/issues/869 */ }
-					<label className="switch">
-						<input type="checkbox" onChange={ handleIconNamesToggle } checked={ showIconNames } />
+					{ /* eslint-disable-next-line jsx-a11y/label-has-associated-control -- https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/issues/578 */ }
+					<label className="switch" htmlFor="showIconNames">
+						<input
+							id="showIconNames"
+							type="checkbox"
+							onChange={ handleIconNamesToggle }
+							checked={ showIconNames }
+						/>
 						<span className="handle"></span>
 						<span className="switch-label" data-on="On" data-off="Off"></span>
 					</label>