+++ projects/plugins/boost/app/lib/Cornerstone_Pages.php
@@ -56,6 +56,7 @@ class Cornerstone_Pages implements Has_Setup {
 		$homepage = array( '' );
 
 		$urls = array_unique( array_merge( $homepage, $woocommerce_pages, $yoast_cornerstone_pages ) );
+		$urls = array_map( 'untrailingslashit', $urls );
 
 		return array_slice( $urls, 0, $max_pages );
 	}
@@ -129,6 +130,7 @@ class Cornerstone_Pages implements Has_Setup {
 		return array(
 			'max_pages'         => $this->get_max_pages(),
 			'max_pages_premium' => static::PREMIUM_MAX_PAGES,
+			'default_pages'     => array_map( 'home_url', $this->default_pages() ),
 		);
 	}