+++ projects/plugins/jetpack/_inc/client/my-plan/my-plan-body.jsx
@@ -310,6 +310,7 @@ class MyPlanBody extends React.Component {
 			case 'is-personal-plan':
 			case 'is-premium-plan':
 			case 'is-jetpack-starter-plan':
+			case 'is-jetpack-growth-plan':
 			case 'is-security-t1-plan':
 			case 'is-security-t2-plan':
 			case 'is-business-plan':